Move parameters and rename variable
