+++ changelog_entry.yaml
@@ -1,0 +1,5 @@
+- bump: patch
+  changes:
+    fixed:
+      - Relocate some ACA parameters to clarify their role.
+      - Rename one ACA variable to clarify its role.
