+++ policyengine_us/parameters/gov/aca/slspc/max_adult_age.yaml
@@ -1,13 +1,15 @@
-description: Maximum age for children under the ACA.
+description: Maximum age for ACA SLSPC amounts.
 metadata:
   period: year
-  label: ACA maximum child age
+  label: ACA maximum SLSPC adult age
   reference:
+    - title: Medicare & the Health Insurance Marketplace
+      href: https://www.medicare.gov/Pubs/pdf/11694-Medicare-and-Marketplace.pdf
     - title: KFF Health Insurance Marketplace Calculator
       href: https://www.kff.org/interactive/subsidy-calculator/
       # https://www.kff.org/interactive/subsidy-calculator-202?/ for prior years
       # The above references have (since 2014) contained a pull-down menu that
       # has this title:
-      #  6. Number of children (20 and younger) enrolling in Marketplace coverage
+      #   5. Number of adults (21 to 64) enrolling in Marketplace coverage
 values:
-  2014-01-01: 20
+  2014-01-01: 64
