+++ policyengine_us/parameters/gov/aca/slspc/max_child_age.yaml
@@ -1,0 +1,13 @@
+description: Maximum age for ACA SLSPC amounts.
+metadata:
+  period: year
+  label: ACA maximum SLSPC child age
+  reference:
+    - title: KFF Health Insurance Marketplace Calculator
+      href: https://www.kff.org/interactive/subsidy-calculator/
+      # https://www.kff.org/interactive/subsidy-calculator-202?/ for prior years
+      # The above references have (since 2014) contained a pull-down menu that
+      # has this title:
+      #  6. Number of children (20 and younger) enrolling in Marketplace coverage
+values:
+  2014-01-01: 20
