+++ policyengine_us/tests/policy/baseline/gov/aca/slspc/person_aca_slspc_ca.yaml
@@ -1,7 +1,7 @@
 - name: person_aca_slspc_ca unit test 1
   period: 2022
   input:
-    aca_trimmed_age: 25
+    aca_slspc_trimmed_age: 25
     is_aca_ptc_eligible_ca: true
     state_fips: 6  # CA
     county_fips: 0  # UNKNOWN
