+++ policyengine_us/variables/gov/aca/eligibility/aca_child_index.py
@@ -9,7 +9,7 @@ class aca_child_index(Variable):
 
     def formula(person, period, parameters):
         age = person("age", period)
-        is_child = age <= parameters(period).gov.aca.max_child_age
+        is_child = age <= parameters(period).gov.aca.slspc.max_child_age
         child_rank = person.get_rank(person.tax_unit, -age, condition=is_child)
         # child_rank of 0 ==> oldest; ties broken by order in the tax unit
         ADULT_VALUE = 99
